update type field stock_quantity

+++ database/migrations/2025_07_09_143214_update_stock_quantity_default_in_products_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->unsignedInteger('stock_quantity')->default(0)->change();
+            $table->decimal('stock_quantity', 10, 2)->default(0)->change();
         });
     }
